+++ src/Client.test.ts
@@ -1,79 +1,73 @@
-import { Report, Components, HybiscusClient } from "./index";
+import { enableFetchMocks } from "jest-fetch-mock";
+enableFetchMocks();
+import fetch from "jest-fetch-mock";
+import { Report, Components, HybiscusClient } from "./";
+const {
+    Core: { Text },
+} = Components;
 
-// eslint-disable-next-line @typescript-eslint/no-var-requires
-jest.mock("cross-fetch", () => require("fetch-mock-jest").sandbox());
-import fetchMock from "cross-fetch";
-
-afterEach(() => {
-    fetchMock.mockReset();
+beforeEach(() => {
+    fetch.resetMocks();
 });
 
-test("Build PDF report", async () => {
-    const apiKey = "P09U8Y7G";
-    const taskID = "apoafajfiqwu38r";
-    const client = new HybiscusClient(apiKey);
-    const postResponse = {
-        task_id: taskID,
-        status: "QUEUED",
-    };
-    const getResponse = {
-        task_id: taskID,
-        status: "SUCCESS",
-    };
-    const report = new Report({
-        reportTitle: "Report title",
-        reportByline: "Report byline",
-    }).addComponent(new Components.Text({ text: "Text component" }));
-    const getURL = `https://api.hybiscus.dev/api/v1/get-task-status?api_key=${apiKey}&task_id=${taskID}`;
-    fetchMock.post(
-        "https://api.hybiscus.dev/api/v1/build-report",
-        postResponse
-    );
-    fetchMock.get(getURL, getResponse);
-    const result = await client.buildReport({ report });
-    expect(fetchMock).toHaveFetched(
-        "https://api.hybiscus.dev/api/v1/build-report"
-    );
-    expect(fetchMock).toHaveFetched(getURL);
-    expect(result).toStrictEqual({
-        taskID: getResponse.task_id,
-        status: "SUCCESS",
-        errorMessage: null,
-        url: `https://api.hybiscus.dev/api/v1/get-report?task_id=${result.taskID}&api_key=${apiKey}`,
+describe("Testing Client", () => {
+    it("Build PDF report", async () => {
+        const apiKey = "P09U8Y7G";
+        const taskID = "apoafajfiqwu38r";
+        const postResponse = {
+            task_id: taskID,
+            status: "QUEUED",
+        };
+        const getResponse = {
+            task_id: taskID,
+            status: "SUCCESS",
+        };
+        fetch
+            .once(JSON.stringify(postResponse))
+            .once(JSON.stringify(getResponse));
+
+        const client = new HybiscusClient(apiKey);
+        const report = new Report({
+            report_title: "Report title",
+            report_byline: "Report byline",
+        }).addComponent(new Text({ text: "Text component" }));
+        const result = await client.buildReport({ report });
+
+        expect(result).toStrictEqual({
+            taskID: getResponse.task_id,
+            status: "SUCCESS",
+            errorMessage: null,
+            url: `https://api.hybiscus.dev/api/v1/get-report?task_id=${result.taskID}&api_key=${apiKey}`,
+        });
     });
-});
 
-test("Preview PDF report", async () => {
-    const apiKey = "P09U8Y7G";
-    const taskID = "apoafajfiqwu38r";
-    const client = new HybiscusClient(apiKey);
-    const postResponse = {
-        task_id: taskID,
-        status: "QUEUED",
-    };
-    const getResponse = {
-        task_id: taskID,
-        status: "SUCCESS",
-    };
-    const report = new Report({
-        reportTitle: "Report title",
-        reportByline: "Report byline",
-    }).addComponent(new Components.Text({ text: "Text component" }));
-    const getURL = `https://api.hybiscus.dev/api/v1/get-task-status?api_key=${apiKey}&task_id=${taskID}`;
-    fetchMock.post(
-        "https://api.hybiscus.dev/api/v1/preview-report",
-        postResponse
-    );
-    fetchMock.get(getURL, getResponse);
-    const result = await client.previewReport({ report });
-    expect(fetchMock).toHaveFetched(
-        "https://api.hybiscus.dev/api/v1/preview-report"
-    );
-    expect(fetchMock).toHaveFetched(getURL);
-    expect(result).toStrictEqual({
-        taskID: getResponse.task_id,
-        status: "SUCCESS",
-        errorMessage: null,
-        url: `https://api.hybiscus.dev/api/v1/get-report?task_id=${result.taskID}&api_key=${apiKey}`,
+    it("Preview PDF report", async () => {
+        const apiKey = "P09U8Y7G";
+        const taskID = "apoafajfiqwu38r";
+        const client = new HybiscusClient(apiKey);
+        const postResponse = {
+            task_id: taskID,
+            status: "QUEUED",
+        };
+        const getResponse = {
+            task_id: taskID,
+            status: "SUCCESS",
+        };
+        fetch
+            .once(JSON.stringify(postResponse))
+            .once(JSON.stringify(getResponse));
+
+        const report = new Report({
+            report_title: "Report title",
+            report_byline: "Report byline",
+        }).addComponent(new Text({ text: "Text component" }));
+        const result = await client.previewReport({ report });
+
+        expect(result).toStrictEqual({
+            taskID: getResponse.task_id,
+            status: "SUCCESS",
+            errorMessage: null,
+            url: `https://api.hybiscus.dev/api/v1/get-report?task_id=${result.taskID}&api_key=${apiKey}`,
+        });
     });
 });