+++ src/Client.ts
@@ -9,7 +9,7 @@ import { HttpTransport, ITaskStatusResponse } from "./HttpTransport";
 interface IPDFReport {
     url: string | null;
     taskID: string | null;
-    status: "CREATED" | "SUCCESS" | "FAILED" | "RUNNING" | "QUEUED";
+    status: "CREATED" | "SUCCESS" | "FAILED" | "RUNNING" | "QUEUED" | null;
     errorMessage: string | null;
 }
 
@@ -54,6 +54,7 @@ class HybiscusClient {
             );
             status = response.status;
             taskID = response.taskID;
+            errorMessage = response.error;
         }
         if (taskID !== null) {
             const response = await this.api.getTaskStatus(taskID);
@@ -64,6 +65,13 @@ class HybiscusClient {
                     url: null,
                     taskID,
                     status: "FAILED",
+                    errorMessage: errorMessage || null,
+                };
+            } else if (status === "SUCCESS") {
+                return {
+                    taskID,
+                    url: `https://api.hybiscus.dev/api/v1/get-report?task_id=${taskID}&api_key=${this.apiKey}`,
+                    status: "SUCCESS",
                     errorMessage: errorMessage || null,
                 };
             }
@@ -90,7 +98,7 @@ class HybiscusClient {
             return {
                 url: null,
                 taskID,
-                status: "FAILED",
+                status,
                 errorMessage: errorMessage || null,
             };
         }
@@ -124,6 +132,7 @@ class HybiscusClient {
             );
             status = response.status;
             taskID = response.taskID;
+            errorMessage = response.error;
         }
         if (taskID !== null) {
             const response = await this.api.getTaskStatus(taskID);
@@ -134,6 +143,13 @@ class HybiscusClient {
                     url: null,
                     taskID,
                     status: "FAILED",
+                    errorMessage: errorMessage || null,
+                };
+            } else if (status === "SUCCESS") {
+                return {
+                    taskID,
+                    url: `https://api.hybiscus.dev/api/v1/get-report?task_id=${taskID}&api_key=${this.apiKey}`,
+                    status: "SUCCESS",
                     errorMessage: errorMessage || null,
                 };
             }
@@ -160,7 +176,7 @@ class HybiscusClient {
             return {
                 url: null,
                 taskID,
-                status: "FAILED",
+                status,
                 errorMessage: errorMessage || null,
             };
         }