+++ src/components/charts/index.ts
@@ -1,4 +1,0 @@
-export { LineChart } from "./Line";
-export { BarChart } from "./Bar";
-export { HorizontalBarChart } from "./HorizontalBar";
-export { DoughnutRing } from "./DoughnutRing";