+++ src/components/core/Table.ts
@@ -1,47 +1,37 @@
-import { Component, IComponentDefinition } from "../base";
+import { Component } from "../base";
+import { WidthType, IOptions } from "../types";
+
+export interface ICellDescription {
+    content: string;
+    col_span: number;
+    col_align: string;
+}
+
+export interface RowCellDescription extends ICellDescription {
+    col_bg: string;
+}
+
+export interface ITableOptions extends IOptions {
+    title: string | null;
+    headings: Array<string | ICellDescription>;
+    width?: WidthType;
+    rows: Array<Array<string | RowCellDescription>>;
+    striped?: boolean;
+    table_border?: boolean;
+    horizontal_margin?: number;
+    vertical_margin?: number;
+    col_border?: boolean;
+    col_align?: string[];
+    col_width?: number[];
+    col_bg?: string[];
+}
 
 class Table extends Component {
-    title: string | null;
-    headings: Array<string>;
-    rows: Array<Array<string>>;
-    striped: boolean;
-
-    /**
-     * Table constructor
-     * @param config Component config
-     * @param config.title Table title
-     * @param config.headings Table headings
-     * @param config.rows Table rows
-     * @param config.striped Whether table is striped
-     */
-    constructor({
-        title,
-        headings,
-        rows,
-        striped
-    }: {
-        title: string,
-        headings: Array<string>,
-        rows: Array<Array<string>>,
-        striped: boolean,
-    }) {
-        super({ type: "Table" });
-        this.title = title;
-        this.headings = headings;
-        this.rows = rows;
-        this.striped = striped;
-    }
-
-    getDefinition(): IComponentDefinition {
-        return {
-            type: this.type,
-            options: {
-                title: this.title,
-                headings: this.headings,
-                rows: this.rows,
-                striped: this.striped,
-            }
-        };
+    constructor(
+        options: ITableOptions = <ITableOptions>{},
+        componentType = "Table"
+    ) {
+        super(options, componentType);
     }
 }